requirements.txt
Django==2.2.2
django-filter==21.1
django-ratelimit==4.1.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
PyJWT==1.7.1
pytz==2024.1
sqlparse==0.5.0
